+++ consensus/src/main/java/org/apache/iotdb/commons/partition/GroupType.java
@@ -17,48 +17,11 @@
  * under the License.
  */
 
-package org.apache.iotdb.consensus.common;
-
-import java.util.Objects;
+package org.apache.iotdb.commons.partition;
 
 // TODO Use a mature IDL framework such as Protobuf to manage this structure
-public class ConsensusGroupId {
-
-  private final GroupType type;
-  private final long id;
-
-  public ConsensusGroupId(GroupType type, long id) {
-    this.type = type;
-    this.id = id;
-  }
-
-  public GroupType getType() {
-    return type;
-  }
-
-  public long getId() {
-    return id;
-  }
-
-  @Override
-  public boolean equals(Object o) {
-    if (this == o) {
-      return true;
-    }
-    if (o == null || getClass() != o.getClass()) {
-      return false;
-    }
-    ConsensusGroupId that = (ConsensusGroupId) o;
-    return id == that.id && Objects.equals(type, that.type);
-  }
-
-  @Override
-  public int hashCode() {
-    return Objects.hash(type, id);
-  }
-
-  @Override
-  public String toString() {
-    return "ConsensusGroupId{" + "type=" + type + ", id=" + id + '}';
-  }
+public enum GroupType {
+  PartitionRegion,
+  DataRegion,
+  SchemaRegion
 }
