+++ consensus/src/main/java/org/apache/iotdb/consensus/common/ConsensusGroup.java
@@ -19,6 +19,8 @@
 
 package org.apache.iotdb.consensus.common;
 
+import org.apache.iotdb.commons.partition.ConsensusGroupId;
+
 import java.util.List;
 import java.util.Objects;
