+++ consensus/src/main/java/org/apache/iotdb/consensus/ratis/RatisConsensus.java
@@ -20,7 +20,7 @@
 package org.apache.iotdb.consensus.ratis;
 
 import org.apache.iotdb.consensus.IConsensus;
-import org.apache.iotdb.consensus.common.ConsensusGroupId;
+import org.apache.iotdb.commons.partition.ConsensusGroupId;
 import org.apache.iotdb.consensus.common.DataSet;
 import org.apache.iotdb.consensus.common.Endpoint;
 import org.apache.iotdb.consensus.common.Peer;
