+++ consensus/src/main/java/org/apache/iotdb/consensus/standalone/StandAloneConsensus.java
@@ -20,7 +20,7 @@
 package org.apache.iotdb.consensus.standalone;
 
 import org.apache.iotdb.consensus.IConsensus;
-import org.apache.iotdb.consensus.common.ConsensusGroupId;
+import org.apache.iotdb.commons.partition.ConsensusGroupId;
 import org.apache.iotdb.consensus.common.DataSet;
 import org.apache.iotdb.consensus.common.Peer;
 import org.apache.iotdb.consensus.common.request.IConsensusRequest;
