+++ consensus/src/test/java/org/apache/iotdb/consensus/ratis/UtilsTest.java
@@ -18,8 +18,8 @@
  */
 package org.apache.iotdb.consensus.ratis;
 
-import org.apache.iotdb.consensus.common.ConsensusGroupId;
-import org.apache.iotdb.consensus.common.GroupType;
+import org.apache.iotdb.commons.partition.ConsensusGroupId;
+import org.apache.iotdb.commons.partition.GroupType;
 
 import org.apache.ratis.protocol.RaftGroupId;
 import org.junit.Assert;
